fix EventTarget tests
