+++ js/globals.ts
@@ -95,6 +95,8 @@ window.EventInit = event.EventInit;
 export type EventInit = event.EventInit;
 window.Event = event.Event;
 export type Event = event.Event;
+window.EventListener = eventTarget.EventListener;
+export type EventListener = eventTarget.EventListener;
 window.EventTarget = eventTarget.EventTarget;
 export type EventTarget = eventTarget.EventTarget;
 window.URL = url.URL;
